// src/app/dashboard/modules/inbox/modules/message-list-view/interfaces/message.interface.ts
export interface IMessageQueryParams {
  search: string;
  type: EMessageType;
  assignedTo: string[];
  propertyManagerId: string[];
  messageStatus: string[];
  taskDeliveryFailIds: string[];
  topicId: string;
  propertyId: string;
  excludeUnHappyPath: boolean;
  excludeConversation: boolean;
  limit: number;
  // same as limit
  // TODO: verify to remove either one
  pageLimit?: number;
  page: number;
  onlyTask: boolean;
  status: TaskStatusType;
  mailBoxId: string;
  currentTaskId?: string;
  isLoading?: boolean;
  labelId?: string;
  reminderType?: string;
  isShowIgnore?: boolean;
  reminderTime?: number;
  timestamp?: number;
  showMessageInTask?: boolean;
}

export interface IMarkAsUnreadResponse {
  id: string;
  isRead: boolean;
}

export interface IFlagUrgentMessageResponse {
  id: string;
  isUrgent: boolean;
}

export enum EMessageType {
  MY_MESSAGES = 'MY_MESSAGES',
  TEAM_MESSAGES = 'TEAM_MESSAGES',
  UNASSIGNED = 'UNASSIGNED'
}
// ...
export interface IPayloadForwardMailBox {
  taskIds: string[];
  mailBoxId: string;
  newMailBoxId: string;
}